Add zooming algorithm to zoom in on bubbles.
This adds a zooming algorithm which can be used to zoom in on bubbles. It also changes the way bubbles were
created initially, to account for links between shared nodes (for example in indels).

There is no algorithm for zooming out yet. Also, right now the tree is created again in the graph view for testing purposes,
because the tree is needed for the bubbling algorithm. Furthermore, the FilterBubbles class does not adhere to
proper coding standards at all, but the only aim of this commit is to make a working zoom, which has been done.