Improved GUI layout
Changing the size of the screen doesn't overlap panes anymore
There is a close button for the bottom graph pane (if it is drawn)
The graph now has a background color which corresponds with the tree color
(coloring in the tree after dragging and dropping still has to be implemented)

The css file has also been updated.
The menu bar and menu items now contain readable text and change background upon hovering.