Added the lineage colors to the graph
Fixes #117

This is a temporary solution. Currently the outer border is the lineage color,
the inner border is de graph color and the node color is the parent/leaf color.
(only the visualization is temporary, the rest works, although the file loading
needs some neater solution when the open file dialog is merged)