Resolve Next to Dev (#156)
* Removed everything after the first '.' when parsing genome names

* Added the lineage colors to the graph

Fixes #117

This is a temporary solution. Currently the outer border is the lineage color,
the inner border is de graph color and the node color is the parent/leaf color.
(only the visualization is temporary, the rest works, although the file loading
needs some neater solution when the open file dialog is merged)

* Updated according to feedback

* Update jdk for travis builds (#138)

Closes #137

Some JavaFX classes are only available in newer versions, so travis needs to update prior building.

* Add test dependencies to core

Core will need them to test upcoming code

* Implement filechooser

Implement filechooser. This filechooser is able to scan a directory
for gfa/nwk/xlsx files to use for the application.

Added open button to the File menu

Also added factories for the graph+tree.
These factories can be used to decouple the way
code obtains its graph/tree.

* Added more information to the description panes of the nodes

TODO: add metadata to leaf nodes.

* Completed the metadata part

All metadata is now shown when you click on a leaf node.

* Fixed string alignment

* Use node objects rather than IDs in in/out edges (#146)

Changed the node IDs which were stored in the in/out edges to actual nodes.
The use of IDs did not have any performance or memory benefits,
but looking up node IDs is very cumbersome.

* Add sprint documents for week 5 and 6 (#147)

Sprint for week 6
Retrospective for week 5

* Start in correct directory when browsing (#149)

fixes #148

* Parse and show date of collection (#144)

closes #143

* Fix tests with new API on next

Changes on dev to the API (saving nodes rather than integers) cause the
tests to break. All integers are replaced by nodes.