Fixed the edge drawing performance
Fixes #182

The edges are now drawn in a Canvas, resulting in lower memory
usage and faster drawing speeds.
The edge width is now approximated, so this no longer is a bottleneck
The 328 graph can now be redrawn at around 60 FPS on my PC.
The lowest FPS rate I got was around 20-25.